requirements.txt
torch
torchvision
open-clip-torch
numpy
opencv-python
Pillow
scikit-learn
tqdm
matplotlib
seaborn
pandas
fastapi
uvicorn
python-multipart
scipy
statsmodels